\name{alqrfe-package}
\alias{alqrfe-package}
\alias{alqrfe}
\docType{package}
\title{\packageTitle{alqrfe}}
\description{\packageDescription{alqrfe}}
\section{Package Content}{\packageIndices{alqrfe}}
\author{\packageAuthor{alqrfe}}
\section{Maintainer}{\packageMaintainer{alqrfe}}
\keyword{package}
