% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQRFE_main_v2.R
\name{bic_hat}
\alias{bic_hat}
\title{Bayesian Information Criteria}
\usage{
bic_hat(res, theta, tau, N, p, inf)
}
\arguments{
\item{res}{Numeric vector, residuals.}

\item{theta}{Numeric vector, parameters.}

\item{tau}{Numeric scalar, the percentile.}

\item{N}{Numeric integer, sample size.}

\item{p}{Numeric integer, parameter length.}

\item{inf}{Numeric, internal small quantity.}
}
\value{
BIC value
}
\description{
Bayesian Information Criteria
}
