% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQRFE_main_v2.R
\name{df_hat}
\alias{df_hat}
\title{degrees of fredom}
\usage{
df_hat(theta, N, p, inf)
}
\arguments{
\item{theta}{Numeric vector, parameters to be test}

\item{N}{Numeric integer, sample size.}

\item{p}{Numeric integer, length of theta.}

\item{inf}{Numeric, internal small quantity.}
}
\value{
degrees of fredom
}
\description{
This function estimates the degrees of fredom
}
