% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQRFE_main_v2.R
\name{f_den}
\alias{f_den}
\title{Kernel density}
\usage{
f_den(x, inf)
}
\arguments{
\item{x}{Numeric vector.}

\item{inf}{Numeric, internal small quantity.}
}
\value{
y vector, kernel density estimation.
}
\description{
Kernel density
}
\examples{
x = rnorm(10)
f_den(x, 0.0001)

}
