% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loss_alqr}
\alias{loss_alqr}
\title{Loss adaptive lasso quantile regression with fixed effects}
\usage{
loss_alqr(theta, x, y, z, tau, n, d, mm, lambda, w)
}
\arguments{
\item{theta}{initial values}

\item{x}{design matrix}

\item{y}{vector output}

\item{z}{incident matrix}

\item{tau}{percentile}

\item{n}{N sample size}

\item{d}{columns of x}

\item{mm}{n columns of z}

\item{lambda}{constriction parameter}

\item{w}{weights}
}
\description{
Loss adaptive lasso quantile regression with fixed effects
}
