% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loss_lqr}
\alias{loss_lqr}
\title{Loss lasso quantile regression with fixed effects}
\usage{
loss_lqr(theta, x, y, z, tau, n, d, mm, lambda)
}
\arguments{
\item{theta}{initial values}

\item{x}{design matrix}

\item{y}{vector output}

\item{z}{incident matrix}

\item{tau}{percentile}

\item{n}{N sample size}

\item{d}{columns of x}

\item{mm}{n columns of z}

\item{lambda}{constriction parameter}
}
\description{
Loss lasso quantile regression with fixed effects
}
