% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loss_qrfe}
\alias{loss_qrfe}
\title{Loss quantile regression with fixed effects}
\usage{
loss_qrfe(theta, x, y, z, tau, n, d, mm)
}
\arguments{
\item{theta}{initial values}

\item{x}{design matrix}

\item{y}{vector output}

\item{z}{incident matrix}

\item{tau}{percentile}

\item{n}{N sample size}

\item{d}{columns of x}

\item{mm}{n columns of z}
}
\description{
Loss quantile regression with fixed effects
}
