% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQRFE_main_v2.R
\name{make_z}
\alias{make_z}
\title{Incident matrix Z}
\usage{
make_z(n, N, id)
}
\arguments{
\item{n}{Numeric integer, number of incidents (subjects, units or individuals).}

\item{N}{Numeric integer, sample size.}

\item{id}{Numeric vector of integer, incident identification.}
}
\value{
Z matrix.
}
\description{
Create an Incident matrix Z
}
