% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQRFE_main_v2.R
\name{optim_lqr}
\alias{optim_lqr}
\title{optim lasso quantile regression with fixed effects}
\usage{
optim_lqr(beta, alpha, x, y, z, tau, N, d, n, ngrid, inf)
}
\arguments{
\item{beta}{Numeric vector, initials values}

\item{alpha}{Numeric vector, initials values}

\item{x}{Numeric matrix, covariates.}

\item{y}{Numeric vector, output.}

\item{z}{Numeric matrix, incidents.}

\item{tau}{Numeric scalar, the percentile.}

\item{N}{Numeric integer, sample size.}

\item{d}{Numeric integer, X number of columns.}

\item{n}{Numeric integer, length of alpha.}

\item{ngrid}{Numeric integer, number of iteractions of BIC.}

\item{inf}{Numeric, internal small quantity.}
}
\value{
parametric vector and residuals
}
\description{
This function solves a lasso quantile regression with fixed effects
}
