% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQRFE_main_v2.R
\name{q_cov}
\alias{q_cov}
\title{Covariance}
\usage{
q_cov(alpha, beta, d, inf, n, N, res, method, tau, X, Z)
}
\arguments{
\item{alpha}{Numeric vector.}

\item{beta}{Numeric vector.}

\item{d}{length of beta.}

\item{inf}{Numeric scalar, internal value, small value.}

\item{n}{length of alpha.}

\item{N}{sample size.}

\item{res}{Numeric vector, residuals.}

\item{method}{Factor, "qr" quantile regression, "qrfe" quantile regression with fixed effects, "lqrfe" Lasso quantile regression with fixed effects, "alqr" adaptive Lasso quantile regression with fixed effects.}

\item{tau}{Numeric, identifies the percentile.}

\item{X}{Numeric matrix, covariates.}

\item{Z}{Numeric matrix, incident matrix.}
}
\description{
Estimate Covariance matrix
}
