% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQRFE_main_v2.R
\name{sgf}
\alias{sgf}
\title{Identify significance}
\usage{
sgf(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
y vector Factor, symbol flag of significant p-values.
}
\description{
Identify significance
}
\examples{
n = 10
pvalue = rgamma(10,1,10)
sgf(pvalue) 
  
}
