\name{alqrfe-package}
\alias{alqrfe-package}
\alias{alqrfe}
\docType{package}
\title{\packageTitle{alqrfe}}
\description{\packageDescription{alqrfe}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{alqrfe}
  \packageIndices{alqrfe}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{alqrfe}
  
  Maintainer: \packageMaintainer{alqrfe}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
